requirements.txt
pytorch-lightning==1.2.10
datasets==1.6.2
transformers==4.5.1
onnxruntime
scikit-learn==0.24.2
wandb
torchmetrics
matplotlib
hydra-core
flask